split enums across domain specific files
