+++ api/models/collection.py
@@ -1,7 +1,7 @@
 from sqlmodel import SQLModel
 from typing import List
 
-from api.common.enums import ResponseView
+from api.common.enums.base_enums import ResponseView
 from api.models.base_response_models import BaseResponseModel
 from api.models.base_row_models import GenericDataModel, RowModel
