+++ api/routers/filer/dependencies/parameters.py
@@ -4,7 +4,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from aiohttp import ClientSession
 
 from api.config.settings import get_settings
-from api.common.enums import ResponseFormat
+from api.common.enums.base_enums import ResponseFormat
 from api.common.formatters import clean, print_enum_values
 from api.dependencies.database import DatabaseSessionManager
 from api.dependencies.http_client import HttpClientSessionManager
