+++ api/routers/filer/routes/metadata.py
@@ -2,7 +2,7 @@ from fastapi import APIRouter, Depends, Query
 from fastapi.exceptions import RequestValidationError
 from typing import Union
 
-from api.common.enums import ResponseContent, ResponseFormat, ResponseView
+from api.common.enums.base_enums import ResponseContent, ResponseFormat, ResponseView
 from api.common.exceptions import RESPONSES
 from api.common.helpers import Parameters, ResponseConfiguration
