+++ api/routers/genomics/queries/search.py
@@ -1,7 +1,7 @@
 from enum import auto
 from typing import List
 
-from api.common.enums import EnumParameter
+from api.common.enums.base_enums import EnumParameter
 from api.models.response_model_properties import QueryDefinition
 from api.models.search import RecordSearchResult
