+++ api/routers/genomics/routes/service.py
@@ -2,7 +2,7 @@ from typing import List
 from fastapi import APIRouter, Depends, Query
 from fastapi.exceptions import RequestValidationError
 
-from api.common.enums import ResponseContent, ResponseFormat, ResponseView
+from api.common.enums.base_enums import ResponseContent, ResponseFormat, ResponseView
 from api.common.exceptions import RESPONSES
 from api.common.helpers import Parameters, ResponseConfiguration
